backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-community==0.3.0
pypdf2==3.0.1
elasticsearch==8.15.0
python-dotenv==1.0.1
requests==2.32.3